+++ spec/common/query_hook_shared_examples.rb
