+++ spec/common/spec_helper.rb
@@ -1,4 +1,4 @@
-require_relative './query_runner_shared_examples'
-require_relative './test_runner_shared_examples'
+require_relative './query_hook_shared_examples'
+require_relative './test_hook_shared_examples'
 require_relative './expectations_hook_shared_examples'
 require_relative './integration_shared_examples'
