+++ spec/common/test_hook_shared_examples.rb
