+++ spec/python2/query_hook_spec.rb
