+++ spec/python3/query_hook_spec.rb
