+++ spec/python3/test_hook_spec.rb
